+++ doc/changelog.rst
@@ -13,6 +13,7 @@ Released the TODO.
     * Support Python 3.4, 3.5
     * Redirect to new urls after updating slug. New settings PAGE_REDIRECT_OLD_SLUG
     * Added get_pages_with_tag templatetag
+    * Added tags in JSON export/import
     * Cleanup
 
 Release 1.8.2